+++ R/mslibrary.R
@@ -396,7 +396,7 @@ setMethod("convertToSuspects", "MSLibrary", function(obj, adduct, spectrumType =
         averagedSpecs <- averageSpectraList(allSpecs, avgSpecParams$clusterMzWindow, avgSpecParams$topMost,
                                             avgSpecParams$minIntensityPre, avgSpecParams$minIntensityPost,
                                             avgSpecParams$relMinAbundance, avgSpecParams$absMinAbundance,
-                                            avgSpecParams$minRelCumIntensity, avgSpecParams$method, FALSE, FALSE, FALSE,
+                                            avgSpecParams$maxRelCumIntensity, avgSpecParams$method, FALSE, FALSE, FALSE,
                                             FALSE)
         names(averagedSpecs) <- names(allSpecs)